% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_solution.R
\name{extract_solution}
\alias{extract_solution}
\title{This function takes the output of the linear program solver and constructs the corresponding policy.}
\usage{
extract_solution(list.of.rectangles.dec, decisions, list.of.rectangles.mtp,
  actions, sln)
}
\description{
This function takes the output of the linear program solver and constructs the corresponding policy.
}
