% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_solver.R
\name{get_default_solver}
\alias{get_default_solver}
\title{Return Default Solver}
\usage{
get_default_solver()
}
\value{
A string of the solver type
}
\description{
Return Default Solver
}
\examples{
get_default_solver()
}
