% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_decision_rectangles.R
\name{refine_decision_rectangles}
\alias{refine_decision_rectangles}
\title{Adaptive Enrichment Design Optimization Using Sparse Linear Programming
Authors: Michael Rosenblum, Ethan Fang, Han Liu}
\usage{
refine_decision_rectangles(
  subpopulation.1.proportion = 0.5,
  stage.1.sample.sizes = c(50, 50),
  stage.2.sample.sizes.per.enrollment.choice = matrix(c(50, 50, 0, 0, 150, 0, 0, 150),
    nrow = 4, ncol = 2, byrow = TRUE, dimnames = list(c(),
    c("Subpopulation1Stage2SampleSize", "Subpopulation2Stage2SampleSize"))),
  discretization.parameter = c(1, 1, 10),
  list.of.rectangles.dec = c(),
  LP.iteration = 1,
  round.each.decision.rectangle.to.integer = FALSE,
  set.rectangles.with.identically.valued.neighbors.and.split.others = TRUE,
  sln
)
}
\arguments{
\item{subpopulation.1.proportion}{Proportion of overall population in subpopulation 1. Must be between 0 and 1.}

\item{stage.1.sample.sizes}{Vector with 2 entries representing stage 1 sample sizes for subpopulations 1 and 2, respectively}

\item{stage.2.sample.sizes.per.enrollment.choice}{Matrix with number.choices.end.of.stage.1 rows and 2 columns, where the (i,j) entry represents the stage 2 sample size under enrollment choice i for subpopulation j.}

\item{discretization.parameter}{vector with 3 elements representing initial discretization of decision region, rejection regions, and grid representing Type I error constraints}

\item{list.of.rectangles.dec}{list of rectangles representing decision region partition, encoded as a list with each element of the list having fields $lower_boundaries (pair of real numbers representing coordinates of lower left corner of rectangle), $upper_boundaries (pair of real numbers representing upper right corner of rectangle), $allowed_decisions (subset of stage.2.sample.sizes.per.enrollment.choice representing which decisions allowed if first stage z-statistics are in corresponding rectangle; default is entire list stage.2.sample.sizes.per.enrollment.choice), $preset_decision (indicator of whether the decision probabilities are hard-coded by the user; default is 0), $d_probs (empty unless $preset_decision==1, in which case it is a vector representing the probabilities of each decision); if list.or.rectangles.dec is empty, then a default partition is used based on discretization.parameter.}

\item{LP.iteration}{positive integer used in file name to store output; can be used to avoid overwriting previous computations}

\item{round.each.decision.rectangle.to.integer}{TRUE/FALSE indicator of whether decision probabilities encoded in list.of.rectangles.dec should be rounded to integer values}

\item{set.rectangles.with.identically.valued.neighbors.and.split.others}{TRUE/FALSE indicator of whether decision probabilities encoded in list.of.rectangles.dec should be modified for use in next iteration}

\item{sln}{solution to linear program computed previously}
}
\value{
4 element list containing optimized designs from four classes (with increasing complexity):
}
\description{
Adaptive Enrichment Design Optimization Using Sparse Linear Programming
Authors: Michael Rosenblum, Ethan Fang, Han Liu
}
